Let `TwoLevelHashTable` choose its bucket count and its bucket hash
The table was fixed at 256 buckets and always routed on the high bits of the hash
a cell is placed by. Both are template parameters now.

`bits_for_bucket = 0` gives a single bucket: routing folds to a constant, the one
sub-table is stored inline, and every operation compiles down to what the
single-level table would do. A caller that wants no partitioning at all can
therefore stay on this template instead of switching to a different type, which is
what the hash join needs to build serially and in parallel through one set of map
types. `NUM_BUCKETS` becomes `numBuckets()` because the count is no longer a
property of the class alone; the call sites are updated with it.

`BucketHash` decouples bucket selection from cell placement. A `FixedHashMap`
places by the key itself, so routing on the high bits of that "hash" would drop a
dense key range into one bucket. `PartitionedFixedHashMap` gives it a routing hash
that hashes the key's cache line instead, which spreads the range and, as a side
effect, keeps two keys sharing a line in the same bucket - so they cannot be
written under two different locks.

The storage behind the buckets becomes a policy. The open-addressing tables keep
one sub-table per bucket as before; a direct-addressed table keeps its single flat
buffer and lets the buckets partition only the locks. That is also why the min/max
iteration bounds have to be switched off while several buckets are being filled,
and `restoreMinMaxOptimization` re-derives them afterwards.

The cell numbering `offsetInternal` produces is the last piece a partitioned build
needs, since the numbers have to stay distinct across buckets. The prefix sums move
behind `computeBucketPrefix` and `offsetInternalUnsafe`, which lets a probe loop
establish them once for a block rather than pay the "already computed" check on
every lookup.

+++ src/Common/HashTable/StringHashTable.h
@@ -248,7 +248,7 @@ protected:
     using Ts = typename SubMaps::Ts;
     using Self = StringHashTable;
 
-    template <typename, typename, size_t>
+    template <typename, typename, Int32>
     friend class TwoLevelStringHashTable;
 
     T0 m0;
